ANDROID: Incremental fs: Build merkle tree when enabling verity
For incfs files that were created without a merkle tree, enabling verity
requires building a merkle tree first. Although this is the same logic
as verity performs, it is not that easy to reconcile the two given that
incfs has the merkle tree potentially when verity is not enabled.

Bug: 160634504.
Test: incfs_test passes.

Signed-off-by: Paul Lawrence <paullawrence@google.com>
Change-Id: Ia15a4051fa3362820846d65859e3af76b77f8cc4
Git-commit: 34c12f7c757df2b362be860e1b12696b28107914 [Applies cleanly]
Git-repo: https://android.googlesource.com/kernel/common/
Signed-off-by: Anjana Hari <quic_ahari@quicinc.com>