ANDROID: Incremental fs: Fix read_log_test which failed sporadically
Test failure was caused by drop_caches sometimes not dropping a
range of sectors in the middle of a random file.
Fix by changing to reading the file before we populate it,
so the cache is not involved. Also by populating it only on the
last test, we test logging on both populated and unpopulated files.

Rewrite the tests with the new test macros, and make the whole test
far more readable.

Test was taking 100 seconds to run, which was half the total runtime
for all the tests. This was due to the sleeps while reading.
Reduce number of such sleeps by a factor of 10,
halving total test runtime.

Also incfs_test would exit with 0 whether tests failed or not. Fix.

Bug: 170104228
Test: Run incfs_test 100 times with -t10 -f4, and then again 40 time
      from end to end

Signed-off-by: Paul Lawrence <paullawrence@google.com>
Change-Id: Iefeca46c2aa02fdf85e7ce4872d9c352f9517f41
Git-commit: cc95fa2a9e6447433ae0ef3911c5634d834f75df
Git-repo: https://android.googlesource.com/kernel/common/
Signed-off-by: Anjana Hari <quic_ahari@quicinc.com>