fs-verity: rename fsverity_signed_digest to fsverity_formatted_digest
The name "struct fsverity_signed_digest" is causing confusion because it
isn't actually a signed digest, but rather it's the way that the digest
is formatted in order to be signed.  Rename it to
"struct fsverity_formatted_digest" to prevent this confusion.

Also update the struct's comment to clarify that it's specific to the
built-in signature verification support and isn't a requirement for all
fs-verity users.

I'll be renaming this struct in fsverity-utils too.

Change-Id: Ia9b23684d010048f935ef350535f770eb69d5bbd
Acked-by: Luca Boccassi <luca.boccassi@microsoft.com>
Link: https://lore.kernel.org/r/20201113211918.71883-3-ebiggers@kernel.org
Signed-off-by: Eric Biggers <ebiggers@google.com>
Git-commit: c9d0dd24e9aa33e765912db6a19ac0d7fd0e8b75 [Applies cleanly]
Git-repo: https://android.googlesource.com/kernel/common/
Signed-off-by: Anjana Hari <quic_ahari@quicinc.com>