defconfig: gen3auto: disable CONFIG_LTO_CLANG
when using sd-llvm compiler, this flag is n.
When using aosp-llvm compiler, this flag is set y.
This config will cause passing more arguments when
"$(call if_changed,link_multi-m)", which cause it exceed
the limitation for call(execve), it will report error
E2BIG(argument list too long). current this issue is
exposed when compling wlan driver, but this is a common
issue for building module.

change align with sd-llvm and avoid passing too much
arguments to call function.

Change-Id: I0101308d316b8bfcd8342680dc5965b7be36d85a
Signed-off-by: Jingxiang Ge <jge@codeaurora.org>