interconnect: qcom: Change QUP vote_scale to 1
The QUP BCMs expect requests to be made in KHz. However, since we
currently scale client requests by 1000, this requires client requests
to be made in MHz. To align with the rest of the non-HLOS images, to
make things more consistent with other BCMs that use KBps, and to allow
for finer-grained frequency requests, change the scaling factor to 1.
This allows the client requests to be made in KHz.

Change-Id: I618f6394778e8f4ab5d6de09ba2cf2ee124122f3
Signed-off-by: Veera Vegivada <vvegivad@codeaurora.org>