arm64/kdump: Don't offline cpu0 during ipi_cpu_crash_stop
Subsystem crash_shutdown handlers seem to require cpu0 to succeed.
Particularly, subsys_crash_shutdown for CDSP gets stuck when
qcom_smem_state_update_bits() on sa8195.

As ipi_cpu_crash_stop() is only sent to cpus when crash kernel is
loaded and panic has occurred. This effectively means that this code
change affects only when kdump is in progress.

This is a workaround patch and proper fix needs to be done.

Change-Id: I4d7a2ba53e296c138113e63daacc046cded73273
Signed-off-by: Akshay Adiga <aadiga@codeaurora.org>