drivers: soc: qcom: Add SLATE PIL driver support
SLATE is a companion chip for Main SoC for low power needs.
Add SLATE PIL driver to provide load, subsystem restart
and ramdump collection support for companion slate chip.

The driver loads the slate firmware into memory and sends
a command to the secure userspace application to initiate
secure transfer of the firmware into internal memory
of SLATE and bring it out of reset. The subsystem restart
depends on couple of GPIOs present between MSM and SLATE.

Change-Id: I3a97ccdb99a9c744fa902614e7298d87e0155ceb
Signed-off-by: Avaneesh Kumar Dwivedi <akdwived@codeaurora.org>