drivers: pci: msm: Add alternate path for PIPE MUXR
Currently there is a limitation on virtual systems where
clk_set_parent for the PIPE MUX is not supported. As a workaround,
the clock space is opened up to the VM and support is added
to switch the source for the PIPE MUX between bi_tcxo and
the external pipe clock.

Change-Id: I92b4871d6dc3406e75f249e2f68be1aa82b465a8
Signed-off-by: Anant Goel <anantg@codeaurora.org>