msm: thermal: Maintain state in the mitigation state for cpu monitor
If KTM emergency cpu frequency or hotplug gets a trip threshold
trigger notification and if the temperature stays the same as the
recent trip threshold, KTM will re-activate the recently triggered
threshold, resulting in back to back interrupts. To avoid this add
support in KTM to maintain the recently triggered threshold state
and then re-active the threshold based on the last threshold trip.

Change-Id: Ia80fc383f682694de61b04439bc46f7f589d0938
Signed-off-by: Manaf Meethalavalappu Pallikunhi <manafm@codeaurora.org>