block: virtio_blk: solve disorder of disk name in async mode
Probe of all virtio_blk devices costs ~100ms and we must set
driver as async to save this ~100ms to improve linux boot
time. But disk name depends on probe order of virtio_blk
device and can be in disorder after driver set as async now.
Usually we can use partition label instead of disk name in
this case as a solution to this kind of problem, but
unfortunately hyp does not support virtualized GPT now, thus
we cannot get partition label in gvm.
Since virtio-mmio probe is sequential now, index of vdev
generated in the probe is in our expected order. Recording
index of the first virtio_blk device created, we can use
vdev index minus this first index when naming disk. This
way, disk name disorder problem can be solved and they are
always in the sequence when disks are created in mmio probe.
Limitation: all virtio_blk devices must be placed together
in hyp gvm linux config file, so the vdev index of
virtio_blk devices are contiguous. This way, the disk names
are contiguous in vda, vdb, vdc, etc.

Change-Id: Ib467f89dadd5b7ec18c6fbff31d3c03c5b5bd384
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>