init: define early init functions for rootfs mounting
Here we define init functions for our first subsystem which
initializes flash controller, filesystem, and starts our
early userspace applications.
And ufs scan runs in sync mode if early feature is enabled
so it runs in the thread for rootfs mounting and we do not
need additional complicated code to do the sync job.

Add fs subsystem downstream driver.

Change-Id: I73b27a86e9c1d366e401f4f478bafa48002c7bce
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>
Signed-off-by: Kaisheng Yang <kaisheng@codeaurora.org>