drivers: cpufreq: Change the priority of notifier block
There is a race seen when msm_thermal driver attempts to
change the cpu frequency during the resume path. In this
scenario, cpufreq is still not aware of the resume (the
msm_cpufreq_pm_event is not yet called).

To fix this issue, change the priority of the
msm_cpufreq_pm_notifier. This will enable cpufreq honor
the frequency update requests in the resume path.

Change-Id: I2780d2c5a2d239646f3f061fda8088c705b178b0
Signed-off-by: Shreyas K K <shrekk@codeaurora.org>
(cherry picked from commit 5e03e2a7822302bcf2c447e2fba2ff3088268ac0)