msm: adsprpc: Disable routing rpc messages to guest OS
While querying fastrpc capabilities the pd type is being set to
1 (guest OS) to forward query to guest OS. This is causing all
subsequent rpc calls to go to guest OS. This is causing issues
for sensor PD daemon because rpc calls for daemon need to go
to sensor PD.

Unset pd type when querying for fastrpc capabilities.
PD type will remain the same as it was set when session started.

Change-Id: I215ec0795078a9dd352e70c87aa02cbd1a34ddb5
Signed-off-by: Edgar Flores <edgarf@codeaurora.org>