iio: qcom-spmi-adc5-gen3: Add change to compensate for Radd
PMIC failure is observed during connection of WDP battery due to
appearance of VBAT voltage levels on Batt-Therm pin and Batt-id pins.
Hence an extra resistor Radd (configurable) is added on Batt-Therm
and Batt-id pins to prevent the damage.

Add support in S/W to compensate for Radd in Battery Temperature
and Batt-id calculations.

Change-Id: I1537267a0da93a29674166593d01a29de403a297
Signed-off-by: Umang Chheda <quic_uchheda@quicinc.com>