pci: msm: Update of de_emphasis setting depending on ntn3 switch version
In NTN3 switch there are i2c writes performed
in order to update some of the settings prior linkup.
These settings are taken care in upcoming versions of ntn3 chips
i.e, above V1 chips.There are some settings still required for NTN3v2
to solve the gen3 AER's. So added a separate switch reg update
sequence for it.

This change is to support update of de_emphasis settings of ntn3 switch
depending on the chip version.Also added a condition to check the
force update of these settings irrespective of the chip version
if force update flag is set by the client in the dtsi.

Change-Id: I0382fc3f857097813982409c3abb5ff611523469
Signed-off-by: Ramya SR <quic_rsr@quicinc.com>