pci: msm: Bringing EP out of reset prior link up
In msm_pcie_enable,EP is brought out of reset
after the link up.This will cause the enumeration
issues with HSP as NTN3 GPIO is connected to WLAN reset.
So moving the part of bringing the EP out of reset
prior link training and after de-emphasis setting.

The reason for moving it after de-emphasis setting is
to ensure WLAN is brought out of reset after all the switch
settings gets applied.

Change-Id: I0efa3bc75bae5625fabbd731050483ac99d567f7
Signed-off-by: Ramya SR <quic_rsr@quicinc.com>