i2c: i2c-msm-v2: Register an interrupt with adding IRQF_NO_SUSPEND
This change registers an interrupt with additional flag IRQF_NO_SUSPEND,
so as to allow system to not disable the interrupt by
suspend_device_irqs().

This is an experimental change for now, as this has risk to get spurious
irq if if there is any other issue in the system.

Change-Id: I5b2eda52e82dc319e4bb74cb44b9c6ae09ad84c9
Signed-off-by: Mukesh Kumar Savaliya <quic_msavaliy@quicinc.com>