Serial: msm_geni_serial: Avoid the Race condition at stop_rx
For some Auto target when the BT SOC is not connected then we are
not getting any response for the BT Host commands and after that
BT host is sending the crash byte and closing the port.
There are some cases when stop rx sequncer is called asynchronously
from runtime suspend and msm_geni_serial_stop_rx function and due
to synchronization is not available, even the stop rx sequencer is
in progress then also it is trying to stop the rx sequencer again, so
to avoid this issue added atomic variable flag.
This way we are ensuring stop_rx from runtime suspend and shutdown
call can't race with each other.
Note that PM callback is independent to the UART framework call.

Change-Id: I30a33d2018d196fb5823ca4ea6d68eed56bb83a7
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>