Revert "soc: qcom: secure_buffer: add API page_accessible"
This reverts commit 724d27beefe988f953810fcd8bcbc90a2c18dc9e.
During hibernation, zram tries to swap out pages to reduce image
size. As part of this, page private field is beign set to point
to 'struct zspage'.

But a hyp-assign is causing the page private being reset to
'flags', which is causing the page to lose its mapping to
'struct zspage', and eventually causing a kernel panic during
hibernation restore. Revert this commit to solve the issue.

Change-Id: I4b238a4f2930b12a190f88409f978bc1b9d75d5f
Signed-off-by: Shreyas K K <quic_shrekk@quicinc.com>