Revert "soc: qcom: secure_buffer: track all hyp based secure pages"
This reverts commit 532b5fece258134d82a71cb0dcd4c9e4ba3e2eee.
During hibernation, zram tries to swap out pages to reduce image
size. As part of this, page private field is beign set to point
to 'struct zspage'.

But a hyp-assign is causing the page private being reset to
'flags', which is causing the page to lose its mapping to
'struct zspage', and eventually causing a kernel panic during
hibernation restore. Revert this commit to solve the issue.

Change-Id: I0ef269139af2b3eb37aa443e9990c88bbc142e6c
Signed-off-by: Shreyas K K <quic_shrekk@quicinc.com>