ext4: fix race writing to an inline_data file while its xattrs are changing.
The location of the system.data extended attribute can change whenever
xattr_sem is not taken.  So we need to recalculate the i_inline_off
field since it mgiht have changed between ext4_write_begin() and
ext4_write_end().

This means that caching i_inline_off is probably not helpful, so in
the long run we should probably get rid of it and shrink the in-memory
ext4 inode slightly, but let's fix the race the simple way for now.

Change-Id: I65f26989087fa7e6a024d3b9ef8064849bb3af45
Git-commit: 9e445093e523f3277081314c864f708fd4bd34aa
Git-repository: https://git.kernel.org/pub/scm/linux/kernel/git/tytso/ext4.git
Cc: stable@kernel.org
Fixes: f19d5870cbf72 ("ext4: add normal write support for inline data")
Reported-by: syzbot+13146364637c7363a7de@syzkaller.appspotmail.com
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Hiren Gohel <quic_vhiren@quicinc.com>
Signed-off-by: Prashanth HN <quic_phn@quicinc.com>