soc: qcom: hab: optimize spinlock pchan_lock to rwlock
Currently pchan_lock is a spinlock lock,
but we don't need a strict spinlock lock.
Because in most cases, what we need is to read
rather than modify the content that needs to be
protected. rwlock is better in this case.
Also fixed a use of vchans_lock in file hab.c,
read_lock is better in this case.

Change-Id: Id172b9090215c00e71bc31b8c57186396b6862fa
Signed-off-by: Zelong Ren <quic_zeloren@quicinc.com>