mmc: core: Handle err variable properly in shutdown path
issue: "mmc0: error -22 during shutdown" logs seen
in shutdown path.

Initially the err variable is initialized with -EINVAL(22)
Before shutting down the eMMC, it make sure that eMMC
should be in active state, in order to do so
 _mmc_resume will be called in mmc_shutdown.

Since the device is already in active state it returns
from the suspend if check and returns -22 to the
mmc_shutdown and treat as error.

Making err variable 0 in the if check will help to
resolve this issue.

Change-Id: I1e34cdce48be3a4db3568cb8442b5aab9110b60c
Signed-off-by: Hiren Gohel <quic_vhiren@quicinc.com>