spi: spi_msm_geni: Check Error condition for DMA mode
Check the proper error condition for DMA mode and according to
that return the success or failure. Currently we are setting
the RX_RFR_WM register before reading the actual value of tx/rx
fifo depth and due to this some garbage value will be set as RFR
value.

Update patch set 19 from source change.

Change-Id: I056991cb27a40d326acf799e370235ba13945503
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>