soc: qcom: hab: decrease dma_buf refcount in unimport
In uhab, the refcount of the dma_buf is decreased by closing the fd
when unimport the buffer.
Unlike the uhab, the refcount of dma_buf is not decreased in khab's
unimport routine. Invoke dma_buf_put to decrease it. Otherwise, memory
leak would be captured due to unreleased dma_buf.

Change-Id: I5ac5f70ef04b6984f627256b1ea34ee037c1216c
Signed-off-by: lixiang <quic_lixian@quicinc.com>