serial: msm_geni_serial: Enable the interrupts after GSI init
Currently we are enabling the interrupt before GSI intialization
and due to this while enabling interrupt we are disabling the
RX_FIFO_WM interrupt if GSI mode is present. But due to the
gsi_mode flag is not set it was not happening. To avoid this
enable the interrupts after GSI initialization sequence is done.

Change-Id: Ib8bbd65e95c8709ede5f87d6d2a51de963cd2ae2
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>