spi: spi-msm-geni: Increase timeout offset from 5 to 250 msec
This change enhances SPI master timeout offset value from 5 msec 250msec
similer to i2c driver. This works fine as long as the system is fine
in the sense no system level latency and all interrupts are delivered
in time. But in some cases, due to other system load, we may face
timeout issue if interrupt is delayed for more than the expected timeout.

Hence settle the offset to little higher value, as 5 msec is very less
in regular scenario also. This change fixes the timeout issues observed
by reported issues on field.

Change-Id: Ie636262bcd2fb81b5baab0d8be5562fce236692e
Signed-off-by: Mukesh Kumar Savaliya <quic_msavaliy@quicinc.com>