rm64: errata: Enable #845719 Errataum for Kryo2xx Silver
Errata#845719 is also applicable for Kryo2xx Silver. Enable
the appropriate entry for it with rAp4 revision.
Please note that default midr_range logic depends on the
less or greather than logic with "min" and "max" range,
assuming that rX where X will be zero only. This is not
true for all the processors and since it is 4-bit field
it can be greater than the the "max" or pY bits.
We are specifying the direct match values like
version 0xA and revision 0x4 instead here to keep the
logic consistent.

Change-Id: I16b0c2106ae649b8a23b7ebb534c967aebd72774
Signed-off-by: Trilok Soni <tsoni@codeaurora.org>
Signed-off-by: Gaurav Kohli <gkohli@codeaurora.org>
[cchidipu@codeaurora.org: resolve trivial merge conflicts]
[cchidipu@codeaurora.org: Update midrs in arm64_workaround_845719_cpus]
Signed-off-by: Chandra Sai Chidipudi <cchidipu@codeaurora.org>
Signed-off-by: Nagireddy Annem <quic_nannem@quicinc.com>