rpmsg: glink: Make wakeup timeout configurable
The Glink slatecom driver can guarantee that the packet gets queued
to the glinkpkt driver but cannot guarantee the client process will
get time to run if auto sleep is enabled. This config will help
mitigate missed packets on systems where auto sleep is too aggressive.

Use the pm_wakeup_ws_event() api so a hard wakeup can be specified.
This will ensure a resume occurs if the data coming in happens while
the device is going into suspend.

Change-Id: I274c856698edc8c87a310a2e1ce547a889368545
Signed-off-by: Vivek Golani <quic_vgolani@quicinc.com>