usb: dwc3: gadget: Clear delay status when gadget driver starts
The gadget driver is not ready to queue the DATA stage TRB for
a three-stage transfer when pullup is disabled and then enabled.
This is due to the timeout that occurred earlier in the STATUS
phase. To fix this, make sure the delayed_status flag is cleared
when the gadget driver starts.

Change-Id: Iab3f5762a421528c0c3da6ef6ac2294e943d6200
Signed-off-by: Selvarasu Ganesan <quic_selvaras@quicinc.com>