msm: mhi: prevent crash in case of invalid MHI controller
In cases when link is not established, MHI driver probe
will fail. In case we try to write SSR sysfs in such a scenario
we will see crash due to invalid MHI controller ptr. This change
fixes it and returns an error in case we don't have a valid MHI
controller.

Change-Id: I5a06336d573e50503666d02ff9ede71c82e894f4
Signed-off-by: Nitesh Gupta <quic_nitegupt@quicinc.com>