virtio_rpmb: Add new device type RPMB_TYPE_VIRTIO.
Add new device type RPMB_TYPE_VIRTIO and use it for the virtio_rpmb
device.
The previously used device type RPMB_TYPE_EMMC caused unwanted settings
of the frame structure member block_count to zero which may be in order
to fulfill the eMCC RPMB specification but which violate the
requirements if the virtio_rpmb specification.
In case the backend physical device is indeed a eMMC which cannot be
known by the driver side it is the duty of the device side to handle the
block_count appropriately.

Issue: QCOM-159
Change-Id: I743f471780b88d2b27d49b90a9ff289270eb81d3
Signed-off-by: linlzhan <quic_linlzhan@quicinc.com>