scsi: ufs-qcom: Fix interconnect sync state issue
This change adds fix for interconnect sync state issue.

sync_state is called only once for a device. It’s called when
all the consumer devices of the device have successfully probed.

If during the first attempt at calling sync_state() for a device,
there are still consumers that haven’t probed successfully, the
sync_state() call is postponed and reattempted in the future only
when one or more consumers of the device probe successfully.

As per sync state rule, the node which is enabled in device tree
of respected driver must probe success, means it consider return
value zero as success.

But UFS probe returning negative value for eMMC boot target. For
that reason sync_state operation will never call for interconnect
device. Hence it is leading to interconnect issue.

Change-Id: Ic716cad1ef700393005677dada1ad633db65b81c
Signed-off-by: Maramaina Naresh <quic_mnaresh@quicinc.com>