msm: mhi_dev: Initialize the snd_cmpl flag before submitting read request
Snd cmpl flag will decide to send completion event for transfers.
Mhi layer can set this flag to '1' whenever drivers needs to
send completion event. In net dev layer snd_cmpl is not initialized
to any value, for first few transfers its value will be 0.
But due to snd_cmpl flag is set by the mhi (mhi sets this flag
when ever the event buffer for the channel reaches 80% of total
size or in wrap around condition) there will be situation that
at certain point of time for each and every transfer driver ends up
sending the completion event as we use same set requests
again and again. This will reduce throughput.

So, initializing the snd_cmpl flag to zero.

Change-Id: I6c65db828e7901f3167b89d49ffab679725b3b35
Signed-off-by: Nitesh Gupta <quic_nitegupt@quicinc.com>