msm: msm-geni-se: Avoid double list add issue in SSR path
Currently while turning on clocks we are calling the add_ab_ib and
rmv_ab_ib. When we receive power on request at that time we will
do the voting without checking if it already done. due to this we
will get double_list_add issue and device will crash, so to avoid
the issue add a check and then perform vote/unvote.

Change-Id: I342e792ca9fcfbdb31a355713bfd0eba0d11c752
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>