spi: spi-msm-geni: Turn on the clocks in SSR up sequence
If SPI client is doing some data transfer and in between that
if we get the SSR up notification then we are turning off the
geni clocks which will create timeout for the ongoing data
transfer. so to avoid this turn on the clocks so next data
transfer will be successful.

Change-Id: I2c5f2f03a717c6c0a815d1a62300b20f33862b89
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>