usb: f_gsi: Check CV2X presence if normal ep used
'commit 87af59be0292 ("usb: f_gsi: Allocating GSI EPs
dynamically") introduced regression while allocating
EPs dynamically. It has a logic to check if CV2X
is part of the composition as part of check it
dereferenced a NULL pointer "opts" when CV2X is
not present.
This changes checks if "opts" is valid based on
it decides CV2X present in the composition.

Change-Id: Idac414e30d287cdb35f0d19155347856f061e349
Signed-off-by: Hiren Gohel <quic_vhiren@quicinc.com>