boot_kpi : Replace place_marker calls with update_marker
Boot KPI driver has a limitation to print a maximum of 73
KPI node entries due to sysfs implementation limitation.
This is impacting the warm boot KPI where we have to try
multiple iterations of suspend and resume and total KPI 
entries exceed 73. To fix this issue replace place_marker
calls with update_marker in the required drivers.

Unlike place_maker API update_marker API will first 
remove the marker and add it again. So the marker comes
at end of the table when we 
cat /sys/kernel/bootkpi/kpi_values .

Change-Id: Iffe8af467e77e947fd358a26a40afec254d0f330
Signed-off-by: Deepak Dimri <quic_ddimri@quicinc.com>