serial: msm_geni_serial: Stop RX sequencer during shutdown path
BT Host crashed but SOC is unaware about this and it is continuously
sending wakeup bytes. Shutdown is called but RX sequencer fails to
stop. Eventually the port is closed but rx sequencer is still active.
When Host opens the port again there is still some data in the RX
queue which leads to Rx framing and break error.

When shutdown is called, don't start the rx sequencer, stop the rx
sequencer so that when Host opens port again there is no data queued
in the RX buffer.

Change-Id: Iff03a805a12c0557f8a398299a36fed82b88865c
Signed-off-by: Prasanna S <quic_prass@quicinc.com>
(cherry picked from commit 4bafe7e01c8825c8e756855c5364e50b32189a12)