qcom_scm: Add support for QCPE 3.0.0 on HOS.
Starting with coqoshv-ext-qcpe-service v3.0.0 the SMC has been implemented
with channel based communication. The major change introduced in this release
is the SMC synchronous exception handling.

The existing solution was based on COQOSHV QCPE Unikraft port that enabled
COQOSHV SEF (Synchronous Exception Forwarding) feature to forward SMC calls
from COQOSHV Guest VM to QCPE VM.
This solution was transparent from virtualization point of view. The drawback
of this solution was the VCPU getting blocked during the SMC call because
it was not scheduled by the hypervisor until SMC return.

A new design called "Channel based SMC" was developed that utilizes COQOSHV
channels, shared memory and interrupts to efficiently handle the SMC calls
from Guest VMs.
Instead of the SMC call, the HVC instruction now triggers the communication
with QCPE and the response is signalled with an interrupt. The difference
is that HVC is used only to signal and does not block the Guest vCPU.

This solution improves the IRQ handling in the Guest VM during the 'SMC' calls.
Previously it was impacted by the VCPU being blocked.

The new design is comparable to QTI's HAB implementation from performance
point of view.

This commit adds the necessary support in the SCM driver.

Changes:

* New module: qcom_scm-coqoshv
* Module initialization triggered at first SMC call
* Optionally allows simultaneous non-atomic SMC calls

New kernel config items:

* CONFIG_QCOM_SCM_QCPE_HOS:
  Enable Channel based SMC in SCM driver
* CONFIG_QCOM_SCM_LOCK_DISABLE:
  Bypass SMC call mutex protection and allows execute
  simultaneus non-atomic SMC calls

Signed-off-by: Robert Kuban <robert.kuban@opensynergy.com>
Signed-off-by: Ulrich Meis <Ulrich.Meis@opensynergy.com>
Signed-off-by: Przemyslaw Marczak <przemyslaw.marczak@opensynergy.com>

Issue: COQOS-16035
Change-Id: I6377106be40fcc0f0c36154caf05ee2f13ddbc1a
(cherry picked from commit 0e5e624eee99c8f6e7cfa7ae3ff9947d9d0ca416)
Signed-off-by: linlzhan <quic_linlzhan@quicinc.com>