power: supply: smb5: Add changes to optimize charger probe time
Add changes to reduce the latency in the probe execution time,
based on a DT property, which will move the execution of all the
deferrable parts of the probe like interrupt registration, sysfs
creation etc, into a scheduled work queue function, to allow
the charger probe to complete sooner.

Also, add DT property "qcom,use-init-work" to enable
this feature.

Change-Id: I19101b316d0f3ccfee75c9509b101073579e92f7
Signed-off-by: Umang Chheda <quic_uchheda@quicinc.com>