usb: f_gsi: Correction while checking if V2X is supported
issue : In SA415 + VT use case when V2X is present in
composition it is mandatory to assign HW EPs to V2X function,
which fails here due to the if check being wrongly added.

This change will check if V2X fs.descriptors are present, based
on it allocation EP happens.

Change-Id: I00b9a7025a985c8c1aff118e4a6d7c580c90fbd9
Signed-off-by: Hiren Gohel <quic_vhiren@quicinc.com>