msm: msm-geni-se: vote ab_noc value same as ib
Currently we are not voting for ab_noc, due to that
aggregate vote is not happening for qup-ddr path.
So set ab_noc value same as ib.

Correct the logic used for comparisons of vote values as well.
Currently, we are voting incorrectly since we are comparing
the Hz value to the Khz value.

Change-Id: I61deb31bbcf2a8e782bf490df35801bca65fd0dc
Signed-off-by: Viken Dadhaniya <quic_vdadhani@quicinc.com>