msm: geni: Enable/disable the SSC corex/core2x clock based on flag
In Hibernation case, multiple times we are trying
to disable/enable SSC corex/2x clocks from client drivers
spi/i2c/uart. Due to this, we are seeing crash.

As we know that SSC corex/2x which is specific for QUP,
not for SEs, so added a new flag whether SSC corex/core2x
enable/disable.

Based on flag we are going to enable and disable SSC corex/2x
clocks.

Change-Id: I5274e9c39fd6b21418a78e38fa5e379d360b75f3
Signed-off-by: Praveen Talari <quic_ptalari@quicinc.com>