scsi: ufs-qcom: call ufs shutdown when do ufs-qcom remove
When the customer do graceful shutdown, the board
will be directly poweroff without shutting down the kernel.
In order to ensure that 2nd ufs works good, we need
to call shutdown manually. So we actually call ufs shutdown
when removing ufs-qcom. This work will be done in the
unrecoverable phase of graceful shutdown, so the ufs-qcom
module will not be re-insmod.

Change-Id: Ib9249654229c9e330e2b7ca9a56ea37bd3b183c5
Signed-off-by: Zelong Ren <quic_zeloren@quicinc.com>