soc: qcom: hab: Introduce an option for FE to make hab open call as uninterruptible
In function hab_open_listen() an option is added by using a flag from user. The flag
HABMM_SOCKET_OPEN_FLAGS_UNINTERRUPTIBLE is introduced, on the basis of it wait_event()
is called.

In a special usecase users are added, deleted and switched through a script in a long
hours stress testing. Where user switching triggers fsync which invokes send_fbe_req and
then habmm_socket_open(). Here habmm_socket_open is interruptible and receiving sigkill which
result in the failure of the fsync/fdatasync operation. So request is to make habmm_socket_open
uninterruptible when called for this use case.

This option is only available for HAB clients in kernel space, and it will be HAB clients
responsibility in kernel space to avoid calling any unexpected uninterruptible habmm_socket_open()
since it is not killable.

Change-Id: Id727624d2f4013086553afcb83a38772691621a6
Signed-off-by: Vijay Raj Hada <quic_vhada@quicinc.com>