msm: msm-geni-se: Enable/disable divider after clocks on/off
In power up sequence we will enable/disable the SE clocks for UART
while loading TZ firmware. As per HPG update we need to turn on/off
the clock divider after/before clocks are on/off resp.

This was missed and due to this we are seeing crash at boot time after
enabling UART node.

Change-Id: I25cc70d096734a67c60d81c3fc29eed45b6c43cb
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>