soc: qcom: hab: move linux-specific implementation of hab_ctx_free out side of the common file
hab_ctx_free() may be called in an atomic context on linux,
so we need put the main work(in hab_ctx_free_fn,which
might sleep on linux) to workqueue. But the above
situation doesn't exist in some OSes, so we move the
linux-specific implementation out side of hab.c(the common
core code).

Change-Id: If8eefc2cf754f8a7418eda7e501878efe3089833
Signed-off-by: Deyan Wang <quic_deyawang@quicinc.com>