ASoC: msm-pcm-q6-v2: Add dsp buf check
Current logic copies user buf size of data
from the avail dsp buf at a given offset.
If this offset returned from DSP in READ_DONE event
goes out of bounds or is corrupted, then it can lead to
out of bounds DSP buffer access, resulting in memory fault.
Fix is to add check for this buf offset, if it is within
the buf size range.

Change-Id: Ib4a55cd9b94e707fd832d25d9809205965486621
Signed-off-by: Ramireddy KrishnaKanth Reddy <quic_ramikris@quicinc.com>