pci: msm: Bringing EP out of reset prior link up
In msm_pcie_enable,EP is brought out of reset
after the link up.This will cause the enumeration
issues with HSP as NTN3 GPIO is connected to WLAN reset.
So moving the part of bringing the EP out of reset
prior link training and after de-emphasis setting.

The reason for moving it after de-emphasis setting is
to ensure WLAN is brought out of reset after all the switch
settings gets applied.

Change-Id: I63cbc1b6124093bf915b9e35cd57050a9ca57059
Signed-off-by: Ramya SR <quic_rsr@quicinc.com>
Signed-off-by: jiangtao li <quic_jiangt@quicinc.com>