xhci: prevent double-fetch of ep_ring in transfer event TRBs
The same values are parsed several times from transfer and event
TRBs by different functions in the same call path, all while processing
one transfer event.

As the TRBs are in DMA memory and can be accessed by the xHC host we want
to avoid this to prevent double-fetch issues.

To resolve this pass the already parsed values to the different functions
in the path of parsing a transfer event.

Change-Id: I6a58d38fa26fc23eb121ffa98aaaaffcd166cbeb
Signed-off-by: Akash Kumar <quic_akakum@quicinc.com>