mmc: sdhci-msm: Disable partial_init and clk-scaling to avoid RED error
As part of partial_init, after tuning when we're enabling CQE (Command
Queue Engine) mode, driver sends CMD44 (QUEUED_TASK_PARAM), and the
response received for this is illegal command error.

This is a temporary fix to avoid RED (Response Error Detected) error until
power and performance impact is analyzed with disabling partial_init and
clk-scaling.

Change-Id: I97833e60399d43a611342ef17a82639c9b7ea5be
Signed-off-by: Sachin Gupta <quic_sachgupt@quicinc.com>