i2c: msm-geni: Added Check for bus recovery disabled
In case if when bus is in bad state for some reason
and bus recovery is disabled we are returning 0 and
initiating the transfer which leads to timeout.

Added check for bus recovery enablement to ensure
if its not enable we will not initate the transfer
and return if IO lines in bad state.

Change-Id: Ie34ac08e63c8b880d27cf443021ba6dde01977a9
Signed-off-by: Panicker Harish <quic_pharish@quicinc.com>